Adopt the cast deny in every file already free of cast findings
`files_without_deny_casts` 1,048 -> 957. Re-measured the tree with one
`--force-warn` pass over the five cast lints (`cast_lossless`,
`cast_possible_truncation`, `cast_possible_wrap`, `cast_sign_loss`,
`ptr_as_ptr`) at `--all-targets`: **15,824 findings in 936 files**, and of
the 1,048 files lacking the deny the histogram is **0: 143 · 1-5: 204 ·
6-20: 415 · 21-100: 285 · >100: 1**. This takes the whole zero tier that can
have it.

Of the 143, **56 are `mod.rs`/`lib.rs`** and a deny there reaches the whole
subtree, so all but three (whose subtrees are also clean, and generated) are
skipped. That leaves 87 leaves + 3 clean-subtree parents, split 55 / 35 by
who writes them.

**Hand-written (54 landed).** `crates/nvim/src/utf8proc/tables.rs` is the
one that could not: it is grandfathered at 12,412 lines and the attribute is
7 more, which the line cap forbids. Its `mod.rs` has 16 findings, so the
module cannot take the deny from above either.

**Generated (37).** Never hand-edited: `tools/apigen` grew
`with_cast_deny`, applied in the `trees` loop next to `with_unsafe_code_allow`
and gated by a new 5th tuple field. It hangs off whichever unsafe attribute
the header already carries -- `#![allow(unsafe_code)]` when the chunk holds
unsafe, `#![forbid(unsafe_code)]` when it does not -- so both header shapes
need one anchor list. Four trees opt in (dispatch tables, Lua binding, option
table, builtin table) plus the two single-file headers (`metadata.rs`,
`cmdtable.rs`); `cmdidx.rs` already had it. The msgpack-RPC wrappers stay
out: they decode 224 argument casts by hand.

The Lua binding tree needed three casts fixed first, since a generated file
cannot carry a hand-written fix:

- `NluaPushFlags` is `c_int`, not `c_uint`. Every consumer already took
  `flags: c_int` and cast the constants back at each use, so the type was
  wrong rather than the four `as c_int`s -- retyping it deletes them and the
  two `PUSH`/`PUSH_SPECIAL` constants stop wrapping.
- `lua_createtable(lstate, 0, BINDINGS.len() as c_int)` is now the count
  written out. apigen knows it; a literal needs no cast.

`just lint` is the net for the whole sweep: a deny on a file that is not
ready fails the build outright, which is the point.

+++ crates/nvim/src/lua/converter/push.rs
@@ -305,7 +305,7 @@ pub unsafe fn nlua_push_typval(lstate: *mut lua_State, tv: *mut TypVal, flags: c
         }
         let mut sink = LuaSink {
             lstate,
-            special: flags & super::kNluaPushSpecial as c_int != 0,
+            special: flags & super::kNluaPushSpecial != 0,
         };
         if !encode_typval(&mut sink, tv, c"nlua_push_typval argument".as_ptr()) {
             return false;
